Enhancement: Smaller main.ts
Deleted the ipcMain functions as they were not used.
Moved the backend function for the app start into backend.ts, which
makes the code more readable and structured.
