+++ electronTS/OmnAIView/package.json
@@ -1,7 +1,7 @@
 {
   "name": "OmnAIView",
   "productName": "omnaiview",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "description": "My Electron application description",
   "main": "dist/main.js",
   "scripts": {
