+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "OmnAIView",
-  "version": "1.1.0",
+  "version": "1.3.0",
   "main": "electronTS/OmnAIView/dist/main.js",
   "scripts": {
     "prestart": "npm --prefix angular-frontend run build",
